Import jpackage incubator module
In order to use jpackage not via the Gradle exec task
but via the java ToolProvider, it's necessary to add the
jpackage's incubator module to the classpath (as long as
jpackage is still in incubator). Due to a bug in the the
OpenJDK 14 implementation, this approach still seems to
require OpenJDK 15.